# client-api/docs/swagger.yaml
definitions:
  types.CreateCustomerRequest:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
    required:
    - email
    - name
    type: object
  types.CustomerDto:
    properties:
      balance:
        type: number
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  types.TransactionRequest:
    properties:
      amount:
        type: number
    required:
    - amount
    type: object
  types.UpdateCustomerRequest:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
    required:
    - email
    - name
    type: object
info:
  contact: {}
  description: API para gerenciar clientes e suas contas
  title: Itau Case API
  version: "1.0"
paths:
  /clientes:
    get:
      consumes:
      - application/json
      description: Endpoint para listar todos os usuários
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Lista todos os usuários
      tags:
      - Clientes
    post:
      consumes:
      - application/json
      description: Endpoint para criar um novo usuário
      parameters:
      - description: Dados do usuário
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/types.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.CustomerDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Cria um novo usuário
      tags:
      - Clientes
  /clientes/{id}:
    delete:
      consumes:
      - application/json
      description: Endpoint para deletar um usuário pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Deleta um usuário pelo ID
      tags:
      - Clientes
    get:
      consumes:
      - application/json
      description: Endpoint para obter um usuário pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.CustomerDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Obtém um usuário pelo ID
      tags:
      - Clientes
    put:
      consumes:
      - application/json
      description: Endpoint para atualizar um usuário existente
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Dados do usuário
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/types.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.CustomerDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Atualiza um usuário existente
      tags:
      - Clientes
  /clientes/{id}/depositar:
    post:
      consumes:
      - application/json
      description: Endpoint para depositar um valor na conta do usuário
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Dados do depósito
        in: body
        name: deposit
        required: true
        schema:
          $ref: '#/definitions/types.TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.CustomerDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Deposita um valor na conta do usuário
      tags:
      - Transações
  /clientes/{id}/sacar:
    post:
      consumes:
      - application/json
      description: Endpoint para sacar um valor da conta do usuário
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Dados do saque
        in: body
        name: withdraw
        required: true
        schema:
          $ref: '#/definitions/types.TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.CustomerDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Saca um valor da conta do usuário
      tags:
      - Transações
  /clientes/{id}/transacoes:
    get:
      consumes:
      - application/json
      description: Endpoint para listar o histórico de transações de um cliente, com
        suporte a paginação via query params `page` e `size`.
      parameters:
      - description: ID do usuário (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: 'Número da página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (default: 10)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retorna metadados de paginação e a lista de transações
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Lista todas as transações de um usuário
      tags:
      - Transações
swagger: "2.0"